# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        example: Error message here
        type: string
    type: object
...
  models.Order:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      order_date:
        example: "2024-01-01T12:00:00Z"
        type: string
      status:
        enum:
        - pending
        - preparing
        - ready
        - delivered
        - cancelled
        example: pending
        type: string
      table_id:
        example: 507f1f77bcf86cd799439012
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - status
    - table_id
    type: object
  models.OrderCreateRequest:
    properties:
      status:
        enum:
        - pending
        - preparing
        - ready
        - delivered
        - cancelled
        example: pending
        type: string
      table_id:
        example: 507f1f77bcf86cd799439012
        type: string
    required:
    - status
    - table_id
    type: object
  models.OrderItem:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      food_id:
        example: 507f1f77bcf86cd799439013
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      order_id:
        example: 507f1f77bcf86cd799439012
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
      unit_price:
        example: 15.99
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - food_id
    - order_id
    - quantity
    - unit_price
    type: object
  models.OrderItemCreateRequest:
    properties:
      food_id:
        example: 507f1f77bcf86cd799439013
        type: string
      order_id:
        example: 507f1f77bcf86cd799439012
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
      unit_price:
        example: 150
        type: number
    required:
    - food_id
    - order_id
    - quantity
    - unit_price
    type: object
  models.OrderItemResponse:
    properties:
      id:
        example: 507f1f77bcf86cd799439017
        type: string
      message:
        example: Order item fetched successfully
        type: string
      order_item:
        $ref: '#/definitions/models.OrderItem'
    type: object
  models.OrderResponse:
    properties:
      id:
        example: 507f1f77bcf86cd799439015
        type: string
      message:
        example: Order fetched successfully
        type: string
      order:
        $ref: '#/definitions/models.Order'
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant.com
    name: API Support
  description: This is a restaurant management system API with user authentication,
    food ordering, and invoice management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
paths:
...
  /orderitems:
    get:
      consumes:
      - application/json
      description: Retrieve order items, optionally filtered by order ID
      parameters:
      - description: Filter by Order ID
        in: query
        name: order_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of order items
          schema:
            items:
              $ref: '#/definitions/models.OrderItemResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Order Items
      tags:
      - OrderItem
    post:
      consumes:
      - application/json
      description: Create a new order item
      parameters:
      - description: Order item details
        in: body
        name: orderitem
        required: true
        schema:
          $ref: '#/definitions/models.OrderItemCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order item created successfully
          schema:
            $ref: '#/definitions/models.OrderItemResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Order Item
      tags:
      - OrderItem
  /orderitems/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order item by ID
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order item deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Order item not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Order Item
      tags:
      - OrderItem
    get:
      consumes:
      - application/json
      description: Retrieve a specific order item by its ID
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order item details
          schema:
            $ref: '#/definitions/models.OrderItemResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Order item not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Order Item by ID
      tags:
      - OrderItem
    put:
      consumes:
      - application/json
      description: Update an existing order item
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated order item details
        in: body
        name: orderitem
        required: true
        schema:
          $ref: '#/definitions/models.OrderItemCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order item updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Order item not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Order Item
      tags:
      - OrderItem
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            items:
              $ref: '#/definitions/models.OrderResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            $ref: '#/definitions/models.OrderResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Order
      tags:
      - Order
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Order
      tags:
      - Order
    get:
      consumes:
      - application/json
      description: Retrieve a specific order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/models.OrderResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Order by ID
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: Update an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Order
      tags:
      - Order
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"